Spring JPA1

+++ hibernate/springData/src/main/java/ru/alexkm07/spdata/repository/ProductRepository.java
@@ -3,5 +3,5 @@ package ru.alexkm07.spdata.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.alexkm07.spdata.model.Product;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
 }
